+++ Sources/MapboxMaps/Style/MapStyle/MapStyle.swift
@@ -58,13 +58,14 @@ import MapboxCoreMaps
     @_documentation(visibility: public)
 #endif
 @_spi(Experimental)
-public struct MapStyle: Equatable {
+public struct MapStyle {
     enum LoadMethod: Equatable {
         case uri(StyleURI)
         case json(String)
     }
     var loadMethod: LoadMethod
     var importConfigurations: [StyleImportConfiguration]
+    var content: (() -> MapStyleContent)?
 
     /// Creates a map style using a Mapbox Style JSON.
     ///
@@ -96,6 +97,13 @@ public struct MapStyle: Equatable {
     public init(uri: StyleURI, importConfigurations: [StyleImportConfiguration] = []) {
         self.loadMethod = .uri(uri)
         self.importConfigurations = importConfigurations
+    }
+
+    /// Add MapStyleContent such as Sources, Layers, and Images to your style at runtime
+    public func callAsFunction(@MapStyleContentBuilder content: @escaping () -> MapStyleContent) -> MapStyle {
+        var newMapStyle = self
+        newMapStyle.content = content
+        return newMapStyle
     }
 
     /// [Mapbox Standard](https://www.mapbox.com/blog/standard-core-style) is a general-purpose style with 3D visualization.