+++ Sources/MapboxMaps/Style/StyleProjection.swift
@@ -1,5 +1,5 @@
 /// Describes the projection used to render the map.
-public struct StyleProjection: Hashable, Codable, StyleEncodable {
+public struct StyleProjection: Hashable, Codable, StyleEncodable, Equatable {
 
     /// The name of the projection.
     public var name: StyleProjectionName
@@ -11,5 +11,12 @@ public struct StyleProjection: Hashable, Codable, StyleEncodable {
 
     internal enum CodingKeys: String, CodingKey {
         case name
+    }
+}
+
+@_spi(Experimental)
+extension StyleProjection: PrimitiveMapStyleContent {
+    func _visit(_ visitor: MapStyleContentVisitor) {
+        visitor.model.projection = self
     }
 }