+++ Tests/MapboxMapsTests/Style/Fixtures/Fixtures.swift
@@ -1,4 +1,5 @@
 import Foundation
+import UIKit
 @_spi(Experimental) @testable import MapboxMaps
 
 internal extension Double {
@@ -24,6 +25,12 @@ internal extension String {
 
     static func testConstantValue() -> String {
         return "test-string"
+    }
+}
+
+internal extension Expression {
+    static func testConstantValue() -> Expression {
+        return Expression(.all)
     }
 }
 
@@ -81,6 +88,18 @@ internal extension Bool {
 
     static func testConstantValue() -> Bool {
         return true
+    }
+}
+
+internal extension UIEdgeInsets {
+    static func testConstantValue() -> UIEdgeInsets {
+        return UIEdgeInsets()
+    }
+}
+
+internal extension Slot {
+    static func testConstantValue() -> Slot {
+        return Slot.init(stringLiteral: "bottom")
     }
 }